data access layer with implicit context
