+++ apps/api-server/src/operation-handlers/programs/get-programs.ts
@@ -1,28 +1,10 @@
 import * as api from '@nawadi/api'
-import { useTransaction } from '@nawadi/core'
-import { schema } from '@nawadi/db'
+import { listPrograms } from '@nawadi/core'
 import * as application from '../../application/index.js'
 
 export function getPrograms(): api.GetProgramsOperationHandler<application.Authentication> {
   return async (incomingRequest, authentication) => {
-    const transaction = useTransaction()
-
-    // we don't need a transaction here, but this illustrates one way of
-    // putting the entire operation in a transaction
-    const programsRows = await transaction
-      .select({
-        id: schema.program.id,
-        title: schema.program.title,
-        handle: schema.program.handle,
-      })
-      .from(schema.program)
-
-    // programsRows could be used directly when the generator properly supports nulls.
-    const programsEntity = programsRows.map((row) => ({
-      id: row.id,
-      title: row.title ?? '', // TODO this need support from the generator
-      handle: row.handle,
-    }))
+    const programsEntity = listPrograms()
 
     return {
       status: 200,
